Handle API connection failures more cleanly
Errors connecting to the TheTVDB API during the ShowDatabase constructor
lead to unrelated errors about _closed not being set when the
ShowDatabase object is garbage collected. Now we set _closed to True
before the constructor starts so we know the destructor doesn't have to
close anything until the connection is started.